requirements.txt
Flask
Flask-SQLAlchemy
Flask-CORS
psycopg2-binary
gunicorn
python-dotenv
# Adicionaremos Flask-Migrate e Flask-JWT-Extended se decidirmos usá-los neste mini-sistema.
# Por enquanto, vamos manter o mínimo.